Rename package 'junit' to 'simple_junit'
